requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
pypdf==5.*
python-dotenv==1.*
requests==2.*
google-generativeai==0.7.*
numpy==2.*
